[TK004] Collapse SafeEnumerationExtensions to thin wrappers over SafeEnumerator
Rewrite SafeEnumerationExtensions to add DirectoryInfo and Directory extension blocks delegating to SafeEnumerator. Delete SafeFileEnumerationExtensions.cs. Fix dead using statements in GlobalUsings, GetDirectoryExtensions, tests, and benchmarks.
