+++ DotExtensions.Benchmarking/Benchmarks/IO/SafeFileEnumerationBenchmarks.cs
@@ -6,7 +6,7 @@ using BenchmarkDotNet.Configs;
 using DotExtensions.Benchmarking;
 using System.Threading;
 using DotExtensions.IO.Directories;
-using DotExtensions.IO.Files;
+using DotExtensions.IO;
 
 namespace DotExtensions.Benchmarking.Benchmarks.IO;
