+++ DotExtensions/System/IO/GlobalUsings.cs
@@ -1,3 +1,2 @@
 ﻿global using System.IO;
 global using DotExtensions.IO.Directories;
-global using DotExtensions.IO.Drives;
